WIP: use bash bind instead of TIOCSTI
There is a bug with TIOCSTI where bash echo is turned off (so stuff you
type isn't printed) if it is used to insert text into a command. As an
alternative, bash can bind macros and key sequences to insert the
results of shist into the prompt. This patch is very much a hack and is
getting very complicated to solve many of the edge cases. It is still
incomplete.